FRIENDS WITH BENEFITS: EFFECTIVENESS OF PEER TEACHING ON IMPROVING THE READING COMPREHENSION SKILLS OF GRADE 11 GAS STUDENTS IN SAN PEDRO NATIONAL HIGH SCHOOL
Abstract

This research study aims to know the effectiveness of peer teaching on improving the reading comprehension skills. The researchers collected data from an experiment with 15 selected respondents from Grade 11 GAS students in San Pedro National High School. This is a quantitative study. The research design used in this study was correlational and experimental in order to determine if peer teaching is helpful and effective on improving reading comprehension skills of the learners after a six-day experiment. The researchers conducted this intervention with the use of self-made and validated pre-test and post- test. The results indicated that there is a significant relationship between peer teaching and reading comprehension. It was also shown that peer teaching has a great impact on learners through the results of the post-test. Therefore, peer teaching has a positive impact on the process of comprehending written text.