STRENGHTENING HONESTY AND TIME CONCIOUSNESS: EFECTIVENESS OF ONLINE PRINTING SERVICE AS MARKETING STRATEGY TO THE LEVEL OF SATISFACTION OF STUDENTS IN SAN PEDRO NATIONAL HIGH SCHOOL
Abstract

This research study examined the effectiveness of online printing service as marketing strategy to the level of satisfaction in terms of honesty and time consciousness. The data were collected through the use of validated survey questionnaire. This research study has 31 respondents who came from the customers who bought service on the business simulation - online printing service. The research design that has been used in this study was experimental to determine the level of satisfaction of students to the online printing service as marketing strategy. Finally, the findings of this study revealed that students are highly in favor of online printing service because it affects them positively. Reasonably, the service improved tardiness and helped to value honesty.