TIKAS PAHINGA: LEVEL OF ANXIETY TOWARDS IMPLEMENTATION OF RESERVED OFFICER TRAINING CORPS (ROTC) IN SENIOR HIGH SCHOOL
Abstract

The researchers conducted this study at San Pedro National High School of San Pedro Hagonoy, Bulacan. This study focused on two major variables which is Reserved Officer Training Corps and level of anxiety towards its implementation. The main objective of this study is to measure the Grade 10 students’ level of anxiety towards having ROTC as their core subject. This study is quantitative. The method used was Quasi Experimental. Their respondents were Grade 10-Jade with the population of 68 and universal sampling was used. The researchers used survey questionnaires with pre-test and post-test. Later on, the researchers found out that there is no significant relationship between ROTC and students’ level of anxiety.